tests: write test for install system

+++ tests/run.py
@@ -86,7 +86,8 @@ def run():
     print()
     print(ansi.green + 'All ' + str(count) + ' tests passed successfuly')
     print('Cleaning up...' + ansi.reset)
-    os.remove('testpkgc-1.0.cati')
+    if os.path.isfile('testpkgc-1.0.cati'):
+        os.remove('testpkgc-1.0.cati')
     shutil.rmtree(Env.base_path_dir)
     Temp.clean()
